menu_service/requirements.txt
Flask
Flask-JWT-Extended
Flask-Cors
Flask-SQLAlchemy
psycopg2-binary
requests